pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wake-build"
version = "0.1.1"
description = "Wake container build orchestrator"
requires-python = ">=3.7"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = ["pyyaml"]

[project.scripts]
wake = "wake_build:main"

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
